Moved tests into folders for better structure.
