+++ Camping-BookingSystemTests/Camping-BookingSystemTests.csproj
@@ -37,10 +37,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="BookingTests\" />
     <Folder Include="CampSiteTests\" />
-    <Folder Include="CampSpotTests\" />
-    <Folder Include="CustomerTests\" />
     <Folder Include="SpotTypeTests\" />
   </ItemGroup>
